requirements.txt
tweepy==4.14.0
requests==2.31.0
pixivpy3==3.7.4
python-dotenv==1.0.0
deepl==1.16.1
